@@ -37,7 +37,7 @@ func (s *Server) Throttler() *ConsoleThrottle {
 
 		s.throttler = newConsoleThrottle(throttles.Lines, period)
 		s.throttler.strike = func() {
-			s.PublishConsoleOutputFromDaemon(fmt.Sprintf("Server is outputting console data too quickly -- throttling..."))
+			s.PublishConsoleOutputFromDaemon("Server is outputting console data too quickly -- throttling...")
 		}
 	})
 	return s.throttler
